Initial commit for supporting oauth, functionality that was added to Roundcube 1.5.
It fixes #58.

A new configuration variable "ifpl_oauth_login_redirect" is added that controls whether persistent_login will hide roundcube's user/password login form so that the persistent_login checkbox is presented next to the button that redirects to an oauth authorization server. You would set this to true INSTEAD OF setting Roundcube's "oauth_login_redirect" to true - when the only desired login choice is via oauth.

This seemed like an acceptable approach but if there are other approaches, please share.

This code is working but there are a couple of unfinished items to complete and more testing is needed.

I refactored some of the original code to better handle the two authentication paths (plain and oauth), so it's a bit difficult to look at a diff an see what changed. Sorry.

Some remaining items to fix and test:

fix any indentation issues in source files
more code comments
update documentation
add javascript updates for skins other than elastic

testing:
- test sql changes in mysql and postgres (testing to date has been with sqlite3)
- pt cookie is updated from all areas of ui when access_token is refreshed
- ensure auto-login fails gracefully when refresh_token is expired
- ensure no regressions (username/password functionality works as before)
- more...

Also, I am not a php expert. I tested on PHP 7.2 and don't know if the code will work on newer versions of php or not.